agent/requirements.txt
# Security Camera Agent — Python dependencies
# Install: pip install -r requirements.txt

# ── Computer vision ────────────────────────────────────────────────────────────
opencv-python>=4.9.0
ultralytics>=8.2.0          # YOLOv8

# ── Backend communication ──────────────────────────────────────────────────────
requests>=2.32.0
websocket-client>=1.7.0     # WebSocket connection to backend

# ── Configuration & utilities ─────────────────────────────────────────────────
python-dotenv>=1.0.0
colorama>=0.4.6             # Colored terminal output (optional)

# ── GUI & system tray (new — required for packaged app) ───────────────────────
customtkinter>=5.2.0        # Modern dark-mode UI for pairing screen
pystray>=0.19.5             # System tray icon (Windows/macOS/Linux)
Pillow>=10.3.0              # Image rendering for tray icon

# ── CUDA / GPU (optional) ─────────────────────────────────────────────────────
# Uncomment the correct line for your GPU to enable hardware acceleration.
# Run: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# torch>=2.3.0
